requirements.txt
pyyaml==5.3
BeautifulSoup4==4.8.2
requests==2.22.0
google-api-python-client